backend/pyproject.toml
[project]
name = "helm-backend"
version = "0.1.0"
description = "Helm — Agentic AI Super App Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.21.0",
    "alembic>=1.15.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=3.2.0,<4.0.0",  # passlib incompatible with bcrypt>=4
    "httpx>=0.28.0",
    "mcp>=1.6.0",
    "apscheduler>=3.11.0",
    "loguru>=0.7.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.20",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "anyio>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["app"]
